// src/lib/client/types/schemas.ts
export interface Community {
  id: CommunityId
  name: string
  title: string
  description?: string
  removed: boolean
  published: string
  updated?: string
  deleted: boolean
  nsfw: boolean
  actor_id: string
  local: boolean
  icon?: string
  banner?: string
  hidden: boolean
  posting_restricted_to_mods: boolean
  instance_id: InstanceId
  visibility: CommunityVisibility
}

export interface CommunityAggregates {
  community_id: CommunityId
  subscribers: number
  posts: number
  comments: number
  published: string
  users_active_day: number
  users_active_week: number
  users_active_month: number
  users_active_half_year: number
  subscribers_local: number
}

export interface CommunityBlockView {
  person: Person
  community: Community
}

export interface CommunityFollowerView {
  community: Community
  follower: Person
}
export type CommunityId = number

export interface CommunityModeratorView {
  community: Community
  moderator: Person
}

export interface CommunityView {
  community: Community
  subscribed: SubscribedType
  blocked: boolean
  counts: CommunityAggregates
  banned_from_community: boolean
}
export type CommunityVisibility = 'Public' | 'LocalOnly'
// ...
export interface ImageDetails {
  link: string
  width: number
  height: number
  content_type: string
}

export interface Instance {
  id: InstanceId
  domain: string
  published: string
  updated?: string
  software?: string
  version?: string
}

export interface InstanceBlockView {
  person: Person
  instance: Instance
  site?: Site
}
export type InstanceId = number

export interface InstanceWithFederationState {
  id: InstanceId
  domain: string
  published: string
  updated?: string
  software?: string
  version?: string
  federation_state?: ReadableFederationState
}

export interface Language {
  id: LanguageId
  code: string
  name: string
}
export type LanguageId = number
// ...
export type ListingType = 'All' | 'Local' | 'Subscribed' | 'ModeratorView'
// ...
export interface LocalSite {
  id: LocalSiteId
  site_id: SiteId
  site_setup: boolean
  enable_downvotes: boolean
  enable_nsfw: boolean
  community_creation_admin_only: boolean
  require_email_verification: boolean
  application_question?: string
  private_instance: boolean
  default_theme: string
  default_post_listing_type: ListingType
  legal_information?: string
  hide_modlog_mod_names: boolean
  application_email_admins: boolean
  slur_filter_regex?: string
  actor_name_max_length: number
  federation_enabled: boolean
  captcha_enabled: boolean
  captcha_difficulty: string
  published: string
  updated?: string
  registration_mode: RegistrationMode
  reports_email_admins: boolean
  federation_signed_fetch: boolean
  default_post_listing_mode: PostListingMode
  default_sort_type: SortType
}
export type LocalSiteId = number
// ...
export interface LocalUser {
  id: LocalUserId
  person_id: PersonId
  email?: string
  show_nsfw: boolean
  theme: string
  default_sort_type: SortType
  default_listing_type: ListingType
  interface_language: string
  show_avatars: boolean
  send_notifications_to_email: boolean
  show_scores: boolean
  show_bot_accounts: boolean
  show_read_posts: boolean
  email_verified: boolean
  accepted_application: boolean
  open_links_in_new_tab: boolean
  blur_nsfw: boolean
  auto_expand: boolean
  infinite_scroll_enabled: boolean
  admin: boolean
  post_listing_mode: PostListingMode
  totp_2fa_enabled: boolean
  enable_keyboard_navigation: boolean
  enable_animated_images: boolean
  collapse_bot_comments: boolean
  last_donation_notification: string
}
export type LocalUserId = number

export interface LocalUserView {
  local_user: LocalUser
  local_user_vote_display_mode: LocalUserVoteDisplayMode
  person: Person
  counts: PersonAggregates
}

export interface LocalUserVoteDisplayMode {
  local_user_id: LocalUserId
  score: boolean
  upvotes: boolean
  downvotes: boolean
  upvote_percentage: boolean
}
// ...
export interface MyUserInfo {
  local_user_view: LocalUserView
  follows: Array<CommunityFollowerView>
  moderates: Array<CommunityModeratorView>
  community_blocks: Array<CommunityBlockView>
  instance_blocks: Array<InstanceBlockView>
  person_blocks: Array<PersonBlockView>
  discussion_languages: Array<LanguageId>
}
export interface OpenGraphData {
  title?: string
  description?: string
  image?: string
  embed_video_url?: string
}
export type PaginationCursor = string
export interface PasswordChangeAfterReset {
  token: string
  password: string
  password_verify: string
}
export interface PasswordReset {
  email: string
}

export interface Person {
  id: PersonId
  name: string
  display_name?: string
  avatar?: string
  banned: boolean
  published: string
  updated?: string
  actor_id: string
  bio?: string
  local: boolean
  banner?: string
  deleted: boolean
  matrix_user_id?: string
  bot_account: boolean
  ban_expires?: string
  instance_id: InstanceId
}

export interface PersonAggregates {
  person_id: PersonId
  post_count: number
  comment_count: number
}

export interface PersonBlockView {
  person: Person
  target: Person
}
export type PersonId = number
// ...
export interface PersonView {
  person: Person
  counts: PersonAggregates
  is_admin: boolean
}

export interface Post {
  id: PostId
  name: string
  url?: string
  body?: string
  creator_id: PersonId
  community_id: CommunityId
  removed: boolean
  locked: boolean
  published: string
  updated?: string
  deleted: boolean
  nsfw: boolean
  embed_title?: string
  embed_description?: string
  thumbnail_url?: string
  ap_id: string
  local: boolean
  embed_video_url?: string
  language_id: LanguageId
  featured_community: boolean
  featured_local: boolean
  url_content_type?: string
  alt_text?: string
}

export interface PostAggregates {
  post_id: PostId
  comments: number
  score: number
  upvotes: number
  downvotes: number
  published: string
  newest_comment_time: string
}
export type PostFeatureType = 'Local' | 'Community'
export type PostId = number
export type PostListingMode = 'List' | 'Card' | 'SmallCard'
// ...
export interface PostView {
  post: Post
  creator: Person
  community: Community
  image_details?: ImageDetails
  creator_banned_from_community: boolean
  banned_from_community: boolean
  creator_is_moderator: boolean
  creator_is_admin: boolean
  counts: PostAggregates
  subscribed: SubscribedType
  saved: boolean
  read: boolean
  hidden: boolean
  creator_blocked: boolean
  my_vote?: number
  unread_comments: number
}
// ...
export type RegistrationMode = 'Closed' | 'RequireApplication' | 'Open'
// ...
export interface Site {
  id: SiteId
  name: string
  sidebar?: string
  published: string
  updated?: string
  icon?: string
  banner?: string
  description?: string
  actor_id: string
  last_refreshed_at: string
  inbox_url: string
  public_key: string
  instance_id: InstanceId
  content_warning?: string
}

export interface SiteAggregates {
  site_id: SiteId
  users: number
  posts: number
  comments: number
  communities: number
  users_active_day: number
  users_active_week: number
  users_active_month: number
  users_active_half_year: number
}
export type SiteId = number

export interface SiteView {
  site: Site
  local_site: LocalSite
  local_site_rate_limit: LocalSiteRateLimit
  counts: SiteAggregates
}
export type SortType =
  | 'Active'
  | 'Hot'
  | 'New'
  | 'Old'
  | 'TopDay'
  | 'TopWeek'
  | 'TopMonth'
  | 'TopYear'
  | 'TopAll'
  | 'MostComments'
  | 'NewComments'
  | 'TopHour'
  | 'TopSixHour'
  | 'TopTwelveHour'
  | 'TopThreeMonths'
  | 'TopSixMonths'
  | 'TopNineMonths'
  | 'Controversial'
  | 'Scaled'
export type SubscribedType = 'Subscribed' | 'NotSubscribed' | 'Pending'